Refactor ldms_ls to print metric units with -l
ldms_ls in this patch is backward compatible. It can connect and display
information of an LDMSD running on an older verion. It displays the
information of metric without a unit the same way as it does without the
patch. If a unit string presents, ldms_ls appends the unit to the metric
name in parentheses.