Resolve `-O2` warnings on `strncpy`
On Cray Linux Environment (CLE) 7.0.UP02, gcc-9 seems to be the
preferred GNU C compiler. When '-O2' is supplied to the CFLAGS,
`-Werror=stringop-truncation` (which is not present in GCC 4.8.5 on
CentOS 7) is also enabled and gives the following warnings in many
places in our source tree (which results in error by `-Werror`):
```
    error: ‘strncpy’ specified bound <NUMBER> equals destination size
```

This patch addresses places where the warnings occur, and provides
length checking if necessary.