Fix secondary group authentication
The following commit:

  5e9d1a080c Remove access check duplicated logic in LDMS and ovis_util

attempted to remove duplicated logic, but it turns out that the logic
in ovis_util/util.c has a one-character bug that was not present in the
implementation in core/ldms_xprt_auth.c.

Fix the call to __uid_gid_check() so we are comparing the user's groups
to the object's gid.