Attempt to correct and improve the ldms comments about the set tree lock
__record_set() grabs the __set_tree_lock using __ldms_set_tree_lock(),
so the code commend is corrected to say that the caller must NOT
be holding the ldms set tree lock when it is called.

__ldms_create_set() uses __record_set(), so likewise the caller must
not be holding the lock.

The lock messages are also updated to say "ldms set tree lock"
rather than just "set tree lock". This is a small attempt to disambiguate
the ldms __set_tree_lock from the ldmsd set_tree_lock. Granted,
it would be a layering violation for ldms to depend on a lock in ldmsd,
but every little bit of disambiguation helps.