Make ldmsd_stream_publish_file wait for a response
Without the change, the ldmsd_stream_publish_file() API may close the
connection before the underlying transport sends all the data, resulting
in the peer would not receive the tail of the stream. With the change,
the API waits for a response before it closes the connection with a
timeout. The timeout is configurable by setting the environment variable
LDMSD_STREAM_ACK_TIMEOUT, and the default value is 2 seconds.