Add store-oriented hashes of schema content to sampler_base.
Store authors often need to know if the schema
of two set instances are equivalent, independent of schema string name.
This removes strongly relying on administrators to invent globally
consistent schema names and provides a fast method (comparing hashes)
of verifying that no conflicts in schema content exist.

As computing a hash of the schema is
an expensive operation, computing it once at set creation minimizes work
globally. This patch defines 8 byte integer hash codes (stored in
set_info key/value pairs as 17 byte strings) for 4 kinds of schema
equivalence check: see sampler_base.h enum ldms_schema_hash_t.

The new sampler_base functions are designed with signatures and argument
handling such that they may be applied to multiple set instances in the same
sampler independently, should the specific sampler need it. This also
allows samplers not derived from sampler_base to apply the hash setting functions.

The update to the ldms_ls man page simply identifies that ldms_ls -v
will show set_info strings, such as the updater hints and these schema hash values.

Creating trees based on the string hash key or the string hash key converted
back to a uint64_t allows rapid determination of which store a set instance
can be safely routed to, for those stores (or storage policies)
requiring matching schema.

With this addition, it becomes trivially easy for a single CSV or SOS store
to be configured which accepts any set instance of any schema and dispatches to the appropriate table/file. (with suitable implementation extensions to the store plugins).

An alternative approach to storing the hashes in set_info is to store them in metadata. This might save transmission bytes at the expense of storage bytes, however.