add cpu_num and more error checking to tx2mon
This records the socket number (cpu_num) in the data set
so it doesn't have to be extracted from the set instance name.
It adds return checking so that config fails if a metric add fails.
It fixes an array overrun that happens and is trapped only when compiled in hardened mode
per ARM/rhel7 RPM build defaults (-Wp,-D_FORTIFY_SOURCE=2 -fstack-protector-strong).
It converts a lingering printf into a log call.