store_csv: Repair constant file rename errors
The store_csv code has a bug that results in chown/chmod error
messages every time a file is renamed. A sanitized version of
the error looks like:

  DEBUG: store_csv: rename_output: rename(A, B)
  ERROR: store_csv: rename_output: unable to chmod(A,660): No such file or directory.
  ERROR: store_csv: rename_output: unable to chown(A, <UID>, <GID>: No such file or directory.
  DEBUG: store_csv: rename_output: rename(A, B)

It _appears_ that the second redundant rename of "A" to "B", even though
the chmod and chown of "A" fail immediately before. However, this is not
the case. In fact only the first rename succeeds as we would expect, and
the error from the second rename is simply ignored. The code comment
explaining the ignored error says:

 "enoent happens if altheader = 0 or typeheader = 0"

So it was recognized that the code was incorrect, but the error was ignored
rather than fixing the problem. Rather than going down the path of ignoring
more errors (from chmod and chown), we attempt to fix the root problem.

The issue seems to be that when altheader=0, the header is not a separate
file, but added as a comment in the main CSV file. However the code redundantly
fills in the same file name for the headerfile rather than recognizing this
fact.

This code has become a bit spaghettified over the years, and probably
need an overhaul to make a fully clean fix.

In the mean time we use a small fix to avoid the incorrect file rename
attempts when altheader is zero and when typeheader is zero.