Fix ldmsctl to provide the correct msg length to response handlers
The patch (5195ba3) that unifies the code to gather records into a whole
LDMSD message introduced the bug. The patch does not set the value of
the msglen variable in the __handle_cmd(). Resulting in __handle_cmd
always pass 0 as the response message length. The response handlers of
the status commands, e.g., prdcr_status, updtr_status, and
daemon_status, use the value provided by the __handle_cmd() when they
parse the string into a JSON object.