Fix buffer reallocation in ldmsd_request
The size provided to the `realloc()` call was not the size of the entire
buffer. Instead, the remaining size (`sz`) was given. This caused a
buffer overrun on `ldmsd` (e.g. when processing `thread_stats` request).
This patch corrects the size of the reallocation to be the size of the
entire buffer (`off + sz`). Also use the 4k-round-up of the required
extension to calculate for the new buffer size instead of extending
`__ALLOC_SZ` at a time (which may not be enough).