Fix occasional SEGEV when ldmsd_controller exited
When ldmsd_controller exit, ldms_xprt was closed but the Python main
thread did not wait for the DISCONNECTED event. The segmentation fault
happened when zap_thread delivered DISCONNECTED event which took the
Python GIL in the callback path, but the Python GIL was already
destroyed because the Python main thread was exiting. This was supposed
not to happen because ldmsd_controller used Xprt in a blocking mode and
Xprt.close() shall timed-wait for the DICONNECTED event (like it did for
blocking `Xprt.connect()`).

This patch fixes the Cython `Xprt.close()` to timed-wait in the blocking
mode and adds the missing `do_quit()` call in ldmsd_controller in the
case of EOF (Ctrl-D) command loop exit.