Fix ldmsd_request:stream_xprt_term()
When LDMS xprt terminated in ldmsd, ldmsd went through the collection
of stream clients to clean up the stream clients associated with the
transport. The collection of clients used `<remote_addr,stream_name>`
for rbt keys. However, on rdma transport, the remote addresses of two
transports from the same remote host got resolved to the same address.
This resulted in premature stream client cleanup due to the address
mixed-up. In addition, on the socket transport, `ldms_xprt_sockaddr()`
call in `stream_xprt_term()` returned ENOTCONN the
`ldms_xprt_sockaddr()` exited immediately, leaking the stream client
and the associated LDMS xprt.

This patch modifies the key to of the stream collections to use
`xprt` pointer instead.

NOTE: This has been tested on rdma and sock transports with GDB to
verify the correct find/removal.