Fix ldmsd message with no attributes
For an ldmsd message that has no attributes (e.g.
LDMSD_FAILOVER_PEERCFG_REQ message), the flag of the message was
incorrectly set to only EOM instead of `SOM|EOM`. In addition, the
message length would be exactly 4 bytes. This patch addresses the
message length assertion and the message flags to correctly handle the
message with no attributes.