Use unique connection identifier in message keys
A [pointer:msg_no] key was being used to match
requests with responses. However, if an ldms transport
were disconnected, and then reconnected, the pointer
could be reused in a subsequent connection resulting
in matching a response with the wrong request. This
change generates a unique uint64_t identifier
that is used as part of the key instead of a pointer
to the transport instance.