Add failover for prdcr_subscribe
Without the patch, the failover pair does not share the prdcr_subscribe
configuration. As a result, the failover buddy will not send a
stream_subscribe request to the producers that it has taken over from
its failed buddy. The patch addresses this, and the failover buddy sends
a stream_subscribe request to the producers it has taken over. If there
is another aggregation level, the patch assumes that the failover buddy
has received a stream_subscibe request from its aggregator before taking
over the failed daemon. In other words, the failover buddy does not
automatically republish the stream to the failed daemon’s aggregator.