Make reqc->xprt point to memory on heap instead of stack
The LDMSD config transport’s memory addresses passed by
ldmsd_process_msg_request()’s callers are on the stack. This is not an
issue if the requests come from configuration files. In contrast, the
memory address of an LDMS config transport could be invalid while the
request context is in the message tree because ldmsd_recv_msg() had been
returned already. Moreover, LDMSD could access the invalid address in
the disconnected path when ldmsd_xprt_term() iterates through the
message tree.