Add a periodic flush interval to storage policies
This change adds a periodic call to a storage plugin's flush method.
Previously, this method was never called. The optional 'flush' keyword
is added to the strgp_add command and is used to specify the flush
interval. If the keyword is not specified, the default interval is zero
and the flush method is never called.