Introduce a dedicated thread to print LDMSD's log messages
Logging to the log file becomes a performance bottleneck when a burst of
errors occur, especially on aggregators. We use event-driven programming
to get rid of the log file's lock. The ldmsd_log() function publishes an
event to a worker that is exclusively responsible for printing data to
the log file. There is only one worker that accesses the log file.