Get rid of the msg_tree lock
The patch uses the ovis_ev library to implement the event-driven
paradigm to get rid of the msg_tree lock. Msg_tree worker exclusively
owns the msg_tree. It is responsible for adding and removing a request
context to and from the msg_tree. The patch also revamped how LDMSD
defer start requests and how it starts the deferred requests.

With the change, the request/response handler functions must put back
the 'create' reference of the request context when the context is not in
used anymore, unless they call ldmsd_send_req_response() function, which
puts back the request context's 'create' reference automatically.

Moreover, the change replaces the ref_count in the request context
structure with the ref structure in the ovis_ref library.