Fix use-after-free in zap_sock xid assertion
The send buffer of the outgoing message was freed immediately after the
message was sent to the underlying socket. However, when zap_sock
received a response message, it accessed the (already freed) send buffer
to check if the transaction ID (xid) of the incoming response matched --
a use after free. This patch adds xid to the outstanding io entry and
set the buffer pointer to NULL when it is freed.