Fix the failover's ping logic
The patch fixes two bugs in the failover’s ping logic. First, the ping
response does not include PEERCFG_ACTIVATED in the flag when the peercfg
has been activated. Second, the ping response handler checks whether the
failover buddy has activated the peercfg or not with the wrong logic. It
verifies whether the buddy’s failover has been configured or not instead
of checking if the buddy has started the peercfg. If the buddy has
started the peercfg, it should not start its own cfg yet.

Commit 305313 removed the line that enables __PEERCFG_ACTIVATED from the
failover's flag. This patch turns the __PEERCFG_ACTIVATED flag on when
at least a producer or an updater has been started. The flag is off when
all peer's cfgobjs have been stopped.