ldmsd to NOT update too close to lookup callback (uGNI work around)
This patch is a work around of the `GNI_RC_INVALID_PARAM` error returned
by `GNI_PostRdma()` when ldmsd schedule the set update (RDMA GET) too
close to the lookup completion of the same set (previous RDMA GET).
So, this patch adds a lookup completion timestamp to the prdcr set and
ldmsd will skip an update if the the update schedule is too close to the
lookup completion time. The *magic* time threshold is 1 sec. With this
patch, we don't see `GNI_RC_INVALID_PARAM` on ugni anymore.

NOTE: `ldms/python/test/lokup-update/test.py` consistently reproduce the
      problem.