Increase ldmsd wait time in github 4.3.3-compat-test
The compatibility test failed due to `MemTotal` being 0. This suggested
that the updater had not updated the set yet when the `ldms_ls` queried
it. This was probably due to over-commited, busy github containers,
since the same test passed on the local docker container. Increasing the
wait time from 3 sec to 10 sec solved the issue.