Remove `LDMS_RBD_*` enum from the check list
`LDMS_RBD_INITIATOR`, `LDMS_RBD_LOCAL` and `LDMS_RBD_TARGET` were used
to describe RBD in the ldms processes. They were not used in the wire.
The RBDs were removed from LDMS, and hence the enumerators shall
be removed from the check list as well (or the check will fail).