Fix SIGSEGV in zap_rdma __ep_flush() path
`__ep_flush()` may be called before cq_channel is created. Hence, it
shall check if `cq_channel` is not NULL before trying to reap the cq
event from the channel in the flush path. In addition, `__ep_flush()`
should reap the event from the channel directly instead of calling
`z_rdma_handle_cq_event()` which will re-arm the channel that we are
flushing.