zap_ugni resource configuration by environment variables
zap_ugni consumed too much resources for message buffers on the sampler
daemon. The message buffers were allocated in a big chunk in advanced to
accomodate a number of future endpoints. This may be good for an
aggregator as a chunk of send/recv memory requires a memory registration
and allocate many of the small chunks may underlying resources for
memory registration. However, enforcing a smaller number of large chunk
waste the memory on the sampler daemon unnecessarily, as it expected
only a few connections, not thousands, at a time.

This patch intoduces environment variables so that users can modify the
chunk size appropriately. The default values favor sampler daemon as
changing sampler daemon in production is harder than modifying the
aggregator configuration as we have many more sampler daemons in
production than the aggregator daemons.

Also update `ldmsd.man` to explain the new zap-ugni-specific environment
variables.