Fix SIGSEGV due to NULL prdcr_set in ldmsd
The producer set may not yet be created when the ldmsd receive
DIR_DELETE from the peer. For example, an L2 aggregator issued a dir
request to an L1 aggregator. The L1 aggregator send DIR_LIST to the L2
aggregator in chunks. In the mean time, a sampler (a peer of L1) may die
and causing the L1 to issue a DIR_DELETE message to the L2 aggregator
that the DIR_LIST chunk of that set may not yet arrived. In this case,
`ldmsd_prdcr_set_find()` will return NULL and cause a segmentation
fault. This patch NULL-check the returned prdcr set before using it in
the DIR_DELETE path.