Get rid of the -H cmdline option
The -n and -H options are somewhat redundant. Without the change, the -H
option defines 'myhostname' which is the producer name of the kernel
metric sets. It is also by the test command, greeting, and the request
'set_route'.

The -n option defines the daemon name which is used by the failover
feature. The daemon name is kept rather than the myhostname because the
default of the daemon name combines both hostname and port, which is
more unique than just the hostname.