Rename ibmad/dcgm plugins to ibmad_sampler/dcgm_sampler
Rename the ibmad plugin to ibmad_sampler and the dcgm
plugin to dcgm_sampler. There are install hooks from the
previous plugin shared object names to the new ones to
aid in the transition. Also to maintain compatibility
though the rename, the schema names remain "ibmad"
and "dcgm" respectively.

Fixes #822