Support LDMSD cmd-line options in configuration files (#826)
* Support LDMSD cmd-line options in configuration files

The values specified at the cmd line take precedence over those in
configuration files.

To specify a cmd-line option in a configuration file, use the command
'option'.

   option ATTR=VALUE ...

Users can see all valid attributes of the 'option' command in
ldmsd_controller.

ldmsd_controller> help option

The supported cmd-line options and the corresponding attribute
names of the option commands are:

-l -- log file path or syslog
 log_file=<path to file|syslog>
-v -- log level
 log_level=<level>
-r -- PID file path
 pid_file=<path to file>
-B -- banner mode
 banner=<mode>
-m -- set memory
 set_memory=<size>
-P -- number of worker threads
 worker_threads=<number of threads>
-n -- daemon name
 daemon_name=<name>
-k -- publish the kernel metrics
 publish_kernel=<true|false>
-s -- path to a text file containing kernle metric values
 kernel_set_path=<path to file>
-H -- host or producer name used by the kerel metric sets
 host_name=<name>
-a -- default authentication method
 default_auth=<auth_method>
-A -- arguments of the default authentication method
 default_auth_args=<attribute-value pairs>

* The -x option is already supported by the 'listen' command.
* Users have two approaches to set the default authentication method.
      option default_auth=<name> default_auth_args=<av-pairs>
or
      auth_add name=DEFAULT plugin=<name> <av-pairs>
* Users have two approaches to set the log level.
      option log_level=<level>
or
      loglevel level=<level>

The 'option' command receives as many attributes as users specified.
For example,
> option log_file=/opt/ovis/var/ldmsd.log
or
> option log_file=/opt/ovis/var/ldmsd.log log_level=ERROR

Users can give multiple 'option' lines in a configuration files. LDMSD
uses only the first value of each attribute and ignore the subsequent
value of the same attribute.
For example,
> option worker_threads=16
> option worker_threads=4
The number of worker threads would be 16.

* Cleanup the cmd-line short options, usage, and the man page

There were short options LDMSD does not support anymore and were in the
list. The usage and the ldmsd man page referred to options that do not
exist anymore. This change cleans up the not-supported options.

* Get rid of the -H cmdline option

The -n and -H options are somewhat redundant. Without the change, the -H
option defines 'myhostname' which is the producer name of the kernel
metric sets. It is also by the test command, greeting, and the request
'set_route'.

The -n option defines the daemon name which is used by the failover
feature. The daemon name is kept rather than the myhostname because the
default of the daemon name combines both hostname and port, which is
more unique than just the hostname.

* Make option use the same syntax as the command line