Make __destroy_set_no_lock() only reset contexts related to the set
The commit fbf21b7 made __destroy_set_no_lock() to reset the set handle
in all transport contexts in the transport's list to NULL, including the
contexts that point to other sets. It breaks the link between the
transport and the other sets. A result is that the transport won't put
back the other sets' references when destroying corresponding contexts.