Change ibmad and dcgm sampler names to match new file names
The new plugin names for ibmad and dcgm are ibmad_sampler
and dcgm_sampler, respectively. Symlinks to the previous
names still exist for the old names to aid transition.

Unfortunately, the API makes it such that only one of these
names (practically speaking) can have working metric
set hints. This can be addressed in the future in a new
plugin API.

We make the plugins report the same name as the new
plugin file name. The plugins probably aren't in too
wide use yet, so we err on the side of making this
work for future users.

Fixes #851