Add transaction timeout in store_sos
The SOS store uses transactions to coordinate with
storage management activities such as roll-over. If
one of these activities fails and leaves the transaction
open, it would hang the SOS store and associated I/O
thread. This changes adds a configurable timeout that
the store will wait to acquire a transaction.