fix compid, producer, log level in lustre dynamic plugins
reduced log level to 'debug' for checks of the top directory
that will be missing when each lustre service is off (an expected condition
in some deployments).

added component_id and producer option processing from attributes
for consistency with other plugins and updated man pages.

for mdt,ost which do not run on compute nodes, leaving out jobid metric.