Add the PRDCR_STREAM_DIR message and handler
Upon receiving a PRDCR_STREAM_DIR request, LDMSD replies with the stream
information of the producers that their names are matched a regular
expression.

A user case is for Maestro to send the request to an aggregator to get
its producers' stream information when Maestro does not have acecss to
sampler daemons.