Fix ldmsd config error but not exit bug
The commit c4f19c6 allocated a new config transport memory because the
`ldmsd_req_ctxt` can later corrupt the transport that was in the stack.
The subsequent request handling call will modify `xprt->rsp_err` if
there is an error. However, it will be modifying the newly allocated
xprt memory, leaving the config transport unchanged. Consequently, the
`rsp_err` in the config transport (on the stack) stayed 0.

This patch addresses the issue by copying the `xprt` content that is
modified by the subsequent config handler call back to the config
transport.