Enable logging for legacy stores where a set schema has different definitions
A storage policy maps to a set schema and assumes all sets of the same schema
have the same metric list. If this is not true, due to misconfiguration or
multiple architectures on the system, LDMSD may crash when it tries to store
sets.

With the patch, LDMSD logs a message when it finds a set that has a different
schema digest from its storage policy's cached digest. Also, LDMSD does not
store such sets. In another word, LDMSD will store sets that a storage policy
has cached its digest.