Correct how LDMS calculate the set size when heap size is given
Without the patch, LDMS does not incorporate the given heap size into
calculating the set size but allocates the set’s heap of the given size. The
wrong set size causes memory corruption in the local daemon. The mismatched
heap size value in the set structure and the set’s meta structures cause LDMS’s
clients, e.g., aggregator, to receive the wrong heap size. Thus, the clients
may create a set with a heap smaller than it should be.