Introduce gnulib related files
This approach both gets us more up-to-date files right now and also
give us a cleaner documented approach for updating an maintaining
those file in the future.

The following command was employed:

  gnulib-tool --no-vc-files --source-base=gnulib --import havelib

gnulib was commit 1de03c03215d9919e6b9d768e86027a0a95ea10c
This is a slightly older to avoid the requirement for automake 1.14
(RHEL 7 is still 1.13.X) in the very latest gnulib.