add syscall_name field to linux_proc_sampler.
This metric tells us what system call we are in, if we are blocked, or if
program is running in user code.
It is less revealing of SP and PC and syscall argument data, but may help
diagnose long hangs in an application.
The system call to name mapping problem is supported by the script
ldms-gen-syscall-map, so that the plugin need not be recompiled across
minor kernel changes.