Fix test_sampler config add_schema num_metrics
`add_schema` config command of `test_sampler` can specify either an
explicit list of metrics or `num_metrics` (just the number of metrics).
In the latter case, the `num_metrics` variable in the code was not set
to the value given by the user. It was uninitialized. This made the
sampler miscalculated the required memory size and stomp on others'
memory (e.g.  corrupting malloc chunks).