convert __dlog gdb-only controls to option -L control
Regardless of message origin (file, controller, maestro),
this enables us to dynamically specify the recording of config requests,
failover processing, status query requests in raw or replayable formats.
It allows absolute and delta time stamps to be recorded.
It distinguishes delayed start instructions from the rest in replay format.
It allows the __dlog messages to go to the default log or to a specific file.
It consolidates __dlog from request and failover into a single function.
The -L option can be used in file or command line, though like other cmdline options if
given in file, it is promoted to the front of the file.