Fix the broken config line parsing logic in ldmsd_request_util.c
ldmsd_parse_config_str() does not add the terminating attribute to the
message before converting the message from the host endian to the
network endian. LDMSD may crash while processing a long config line.

The change also addresses the logic that chunks the messages into
multiple records if the message is larger than the maximum of the
transport message size. Before the change, the logic assumes that the
message does not contain the terminating attribute.