Implement the ibmad_records_sampler plugin
Implement the ibmad_records_sampler plugin. Versus the
ibmad_sampler, the major change is that we use a single
list of records per ib interface rather than a metric set
per ib interfaces.

There are a number of other changes and cleanups that
fell out of that change. For instance, we can fully use
the sampler_base code now that we only have a single
metric set.