infiniband: update autoconf tests
Update the inifinband autoconf tests by pulling the
library and header checks out from under the AC_ARG_ENABLE()
for infiniband samplers. This easier allows the results of
the checks to be used by multiple  plugins, and allows
the individual users of the libraries to decide what
constitutes a missing dependency error under what conditions.