Rename the component_id and job_id metrics in slurm_sampler
The 'component_id' metric, an uint64, is the component ID corresponding
to the set. The 'job_id' metric, an uint64, is the job ID corresponding
to the set at a timestamp. However, the 'component_id' metric in
slurm_sampler , an uint64 array, contains the component IDs of each job
in the job_id metric. Similarly, the 'job_id' metric in slurm_sampler,
an uint64 array, contains the slurm job IDs on the system.

store_sos assumes that component_id and job_id are uint64 so that it can
create the joined indices of component_id, job_id, and timestamps.
store_sos hits an assert when it stores the slurm set because the
component_id and job_id are not uint64.

The patch renames slurm_sampler's component_id and job_id metrics
because they have different meanings from the other sampler plugins.