Make store_csv not support multiple lists w/o decomposition
In the case of sets containing multiple lists, store_csv in absense of
the decomposition repeats the last value of the shorter lists for as
many rows as the longest list. The repeated values make the output data
confusing.

The change makes store_csv check whether a set contains more than one
list. If the set contains multiple lists, store_csv reports an error.
The users must stop and delete the storage policy and then create a new
storage policy using a decomposition to store such sets.