Make store_sos not support multiple lists w/o decomposition
In the case of sets containing mulitple lists, store_sos repeats the
last value of the shorter lists for as many rows as the longest list in
absence of the decomposition. The repeated values make the output data
confusing.

The change makes store_sos check whether a set contains more than one
list. If it is the case, store_sos reports an error. The users must stop
and delete the storage policy and then create a new one using a
decomposition to store such sets.