Fix `ldms_xprt_listen_by_name()` for wildcard address
When the caller provides '*' wild card address to
`ldms_xprt_listen_by_name()`, the expected behavior is to listen to all
addresses (0.0.0.0). However, because the AI_PASSIVE has not been
specified in the `ai_flags` hint, '*' is resolved to `127.0.0.1`
instead of `0.0.0.0`. This patch specified `AI_PASSIVE` in
`hint.ai_flags` to fix the issue.