ldms_xprt_new_with_auth() returns a rail of 1 xprt
Make `ldms_xprt_new_with_auth()` to return a rail of single transport.
This is so that the application will be able to use new features that
depends on the new rail transport.

This patch also modify places where `ldmsd` application directly access
fields in the ldms_xprt structure to make it access those information
via an interface instead. The only field that is still left untouched is
the 'x->disconnected' field.

NOTE: This patch passed the followin test cases:
- agg_test
- agg_slurm_test
- ldms_rail_test
- ldms_stream_test