Fix the greeting handler() in the 'path' mode
The path mode violates the null terminating character LDMSD's
request/response protocal.

The path mode forwards the request to the first producer, waits for the
response, then constructs the aggregated information, and replies back
to the client.

The greeting command was introduced to test the request/response logic
in the ldmsd.