[WIP] Fix the delete_proc
The patch is to fix the symptom that sometimes a set of sets are
permanently stuck in the delete tree on the server side. It seems the
server has never received the SET_DELETE reply from the client while the
connection is still connected.

  The delete_proc() skips the sets that there is a transport in the
lookup collection that is connected based on the assumption that 1) the
server had sent the SET_DELETE request to the peers, and 2) the set will
be freed when the server receives the SET_DELETE reply from its peers.

  The patch consults the transports in the lookup collection if they
have the SET_DELETE context for the set or not. If none of the
transports have the SET_DELETE context corresponding to the set, we will
free the set. The logic is to handle when the above assumptions are not
true. We believe this might be a corner case, e.g., we don't properly
handle an error, or the underlying transport stuck in a stale stage.

  A problem that may arise with the patch is the deadlock between the set
lock and the transport lock. We need to take the transport lock inside
the set lock to iterate through the transport's context list.