Change munge error message to info
The munge authentication message consists of the
local and remote addresses of the transport between
the two peers. This message presumes that both peers see
the same addresses (albeit swapped). This assumption is
false in many routed and NAT'd environments.

This change replaces the error with an informational message
so that administrators will be aware of this difference
and can make a determination whether or not this difference
is expected.