Fix ldmsd_stream_subscribe intermingled events output
Multiple threads in `ldmsd_stream_subscribe` can race to handle stream
recv events. The recv handler called `msglog()` multiple times to print
out the EVENT. This lead to intermingled events in the output. This
patch modifies the handler to print the events in single msglog call to
avoid the intermingled output.