Make json_stream_sampler handle UID, GID, and perm
Without the patch, the json_stream_sampler ignores the UID, GID, and
perm given at the config line.

With the patch, the plugin creates LDMS sets with the specified UID,
GID, and permission. If one of the values are not given, it will use the
security data from the received stream data.

After the set has been created, the configuration command set_sec_mod
must be used to modify the sets' security (UID, GID, and perm). The
plugin does not automatically revise the set security according to the
stream data's security changes.